requirements.txt
streamlit>=1.32.0
groq>=0.4.0
python-dotenv>=1.0.0
cryptography>=42.0.0
